// angular-front-end/src/app/models/cliente.ts
export interface ClienteModel {
  id?: Number;
  nome: String;
  email: String;
  dataNascimento: String;
  cep: String;
  logradouro: String;
  complemento: String;
  numero: String;
  bairro: String;
  cidade: String;
  uf: String;
  codigoIbge: String;
  ddd: String;
  telefone: String;
  ativo: Boolean;
}